requirements.txt

celery==5.2.3
Django==4.0.5
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-timezone-field==5.0
djangorestframework==3.13.1
python-crontab==2.6.0
redis==4.3.4
SQLAlchemy==1.4.36
sqlparse==0.4.2
isodate==0.6.1
requests==2.27.1
